+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - The `OnClear` event in RumorState has been changed to use a `ClearType` enum,
   which specifies if everything, just choices, or just dialog was cleared
+- The Pause command will now wait for an advance if the time is less than or
+  equal to 0.
 
 
 ## [0.2.1] - 2016-11-14